rag/requirements.txt
mcp>=0.9.0
ollama>=0.1.0
faiss-cpu>=1.7.4
watchdog>=3.0.0
numpy>=1.24.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
